requirements.txt
# ISR Platform Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
geoalchemy2>=0.14.0
redis>=5.0.0
aiokafka>=0.10.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.1.0
shapely>=2.0.0
numpy>=1.26.0
pandas>=2.1.0

# Database migrations
alembic>=1.13.0

# Machine Learning & Deep Learning
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.25.0
scikit-learn>=1.3.0

# NLP Libraries
spacy>=3.7.0
nltk>=3.8.0

# Model Optimization
optimum>=1.16.0
onnxruntime>=1.16.0

# Satellite Imagery & Geospatial Analysis
earthengine-api>=0.1.384
sentinelsat>=1.2.1
rasterio>=1.3.9
geopandas>=0.14.0
folium>=0.15.0
matplotlib>=3.8.0
pillow>=10.1.0
opencv-python>=4.8.0
pyproj>=3.6.0

# LLM Integration
openai>=1.10.0
anthropic>=0.18.0
tiktoken>=0.5.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.8.0
